requirements.txt
fastapi
uvicorn
joblib
scikit-learn
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]
python-multipart